NF: ``__name__`` is always a reference
This commit is the first commit of what I plan to be a sequence of commit whose
goal is to improve the documentation usability. It's goal is to be an example,
concentrated on ``__name__``, that will allow me to get feedback and show my
goals before working on other missing references.

Whatever the place in the documentation, it may always save a little bit of time
that references are clicable. This allows the reader to easily get more
information if needed. My goal is to go through the part of the documentation in
/Doc and ensure that as many method, variable, members, functions,
classes... becomes links. Actually, I got the idea from PR #29183, where
terryjreedy asked me to add a link to a method in a part of the documentation my
PR was touching. I believe the idea was good and should be systematized.

Since this may be a big change, here are the steps to ensure that it gets easy
to review:
* each commit consider a single word. In this case ``__name__``, so the commit
is really easy (even if boring) to review
* only the lines containing the changed words are changed. No other reformating
occurs. This allow any diff tool to really easily show that no other differences
are introduced.

The downside is that some lines are now longer than 80 characters. However, I
should note that this constraint was already not respected in a lot of place of
the documentation. And until there is an automated formatting tool, I think the
cost is quite small compared to the improvement in the documentation
usability. If those 80 characters are considered a blocker, it would be easy to
reformat all lines in a last commit, once the link creations are done.