bpo-46182: super with one argument clarification
Currently, super with one argument is not clear. Indeed, it still refers to the
notion of "unbound" which was otherwise deleted from the language in Python 3.0
according to its change notes.

This means that this documentation was probably useful for no-one. Indeed:

Either you want to learn more about `super` and you find a case that you don't
understand, you don't even know how it can be useful and whether you should
learn it. The end of this paragraph indicates that you can safely ignore it.

Or you found a codebase using `super` with a single argument and you want to
understand the code you are reading; in which case the documentation was also
useless. While my PR is far shorter and less detailled than
https://stackoverflow.com/questions/30190185/how-to-use-super-with-one-argument
I believe it may help a user who had to understand what this super object do.